review 서버 통신 완료

+++ app/src/main/java/com/syh4834/chabak/api/data/PlaceDetailData.java
@@ -16,7 +16,7 @@ public class PlaceDetailData {
 
     private String placeAddress;
 
-    private int placeStar;
+    private double placeStar;
 
     private double placeLatitude;
